# api/openapi/goeland.swagger.yaml
swagger: "2.0"
info:
  title: goeland/v1/core.proto
  version: version not set
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/documents:
    post:
      summary: |-
        Create / register document metadata. SubjectRef + RecordMetadata are created via Core internally.
        Requires goeland:write; subject, governance, document, version 1, optional
        links and the DOCUMENT_CREATED audit event are written in one transaction.
        Content already held by a live document reuses it (DOCUMENT_REUSED, reused = true).
      operationId: CreateDocument
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1CreateDocumentResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: body
          description: |-
            CreateDocumentRequest registers a document (metadata first; the bytes are
            uploaded beforehand and referenced by content_blob_id). When the content is
            already held by a live document, that document is reused instead (see
            CreateDocumentResponse.reused).
          in: body
          required: true
          schema:
            $ref: '#/definitions/v1CreateDocumentRequest'
      tags:
        - DocumentService
...
  /api/documents/{documentId}/versions:
    get:
      summary: List the versions of a document, newest first. Requires goeland:read.
      operationId: ListDocumentVersions
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1ListDocumentVersionsResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: documentId
          description: document_id is the document; NOT_FOUND when unknown.
          in: path
          required: true
          type: string
      tags:
        - DocumentService
    post:
      summary: |-
        Append a new current version to a document (optionally with new content).
        Requires goeland:write; writes a DOCUMENT_VERSION_ADDED audit event;
        FAILED_PRECONDITION when the document is locked or deleted.
      operationId: AddDocumentVersion
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/v1AddDocumentVersionResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      parameters:
        - name: documentId
          description: document_id is the document; a locked or deleted document is rejected.
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/DocumentServiceAddDocumentVersionBody'
      tags:
        - DocumentService
...
definitions:
...
  DocumentServiceAddDocumentVersionBody:
    type: object
    properties:
      contentBlobId:
        type: string
        description: |-
          content_blob_id is the content from POST /api/documents/upload; empty for a
          metadata-only version. It may repeat an earlier version's content.
      isFinal:
        type: boolean
        description: is_final creates the version already final.
      isRecord:
        type: boolean
        description: is_record declares the version a record, which also makes it final.
      pageCount:
        type: integer
        format: int32
        description: page_count is the number of pages.
      metadata:
        type: object
        description: metadata holds secondary data about the version.
      reason:
        type: string
        description: reason is the justification recorded on the audit event (at most 2000 characters).
    description: AddDocumentVersionRequest appends a new current version to a document.
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
...
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
...
  v1AddDocumentVersionResponse:
    type: object
    properties:
      document:
        $ref: '#/definitions/v1Document'
        description: document is the document; its current_version is the new version.
      version:
        $ref: '#/definitions/v1DocumentVersion'
        description: version is the new version.
      auditEvent:
        $ref: '#/definitions/v1AuditEvent'
        description: audit_event is the DOCUMENT_VERSION_ADDED audit event.
    description: AddDocumentVersionResponse returns the document with its new current version.
...
  v1AuditEvent:
    type: object
    properties:
      id:
        type: string
        description: id is the server-assigned event UUID.
      subjectId:
        type: string
        description: subject_id is the subject the event is about.
      eventType:
        type: string
        description: |-
          event_type is an upper-case code (1-100 characters), e.g. DOCUMENT_CREATED,
          RELATIONSHIP_LINKED or DOCUMENT_FINALIZED.
      actorUserId:
        type: string
        description: |-
          actor_user_id is the authenticated operator who caused the event (not a
          domain ACTOR subject), derived server-side.
      occurredAt:
        type: string
        format: date-time
        description: occurred_at is the server time of the event.
        readOnly: true
      beforeState:
        type: object
        description: |-
          before_state is a snapshot of the changed critical fields before the
          mutation; absent for creations.
      afterState:
        type: object
        description: after_state is a snapshot of the changed critical fields after the mutation.
      reason:
        type: string
        description: reason is the operator-supplied justification (at most 2000 characters).
      correlationId:
        type: string
        description: |-
          correlation_id groups the events of one multi-subject action; empty when
          the event stands alone.
      requestId:
        type: string
        description: request_id is the HTTP request identifier the event was written under.
      metadata:
        type: object
        description: metadata is secondary context for the event.
    description: |-
      AuditEvent is an append-only probative audit record, written on every mutation
      in the same transaction as the change it records.
    required:
      - eventType
...
  v1CreateDocumentRequest:
    type: object
    properties:
      documentTypeCode:
        type: string
        description: document_type_code is the required code of an existing document type.
      title:
        type: string
        description: title is the required title (1-500 characters); surrounding whitespace is trimmed.
      description:
        type: string
        description: description is optional free text (at most 4000 characters).
      officialDate:
        type: string
        description: official_date is the optional legal date as YYYY-MM-DD.
      externalSystem:
        type: string
        description: external_system names the external system of record, if any.
      externalId:
        type: string
        description: external_id identifies the document inside external_system.
      externalUrl:
        type: string
        description: external_url links to the external document; it becomes the subject's canonical URL.
      isFinal:
        type: boolean
        description: is_final creates version 1 already final (document status FINAL).
      isRecord:
        type: boolean
        description: is_record declares version 1 a record, which also makes it final.
      language:
        type: string
        description: language is the ISO 639 content language.
      pageCount:
        type: integer
        format: int32
        description: page_count is the number of pages of version 1.
      metadata:
        type: object
        description: metadata holds secondary extension data.
      initialGovernance:
        $ref: '#/definitions/v1RecordMetadata'
        description: |-
          initial_governance optionally sets the owner, confidentiality and records
          fields; the owner defaults to the operator.

          NOTE: the acting operator is derived server-side from the authenticated
          principal (never trusted from the client). A document's AUTHOR is a separate,
          external ACTOR subject linked via DOCUMENT_AUTHORED_BY_ACTOR — not this operator.
      linkToCaseId:
        type: string
        description: |-
          link_to_case_id is an optional convenience: link the new (or reused)
          document to this case with CASE_HAS_DOCUMENT in the same transaction.
      contentBlobId:
        type: string
        description: |-
          content_blob_id is the content returned by POST /api/documents/upload;
          empty for a metadata-only document or an external reference. NOT_FOUND
          when unknown.
    description: |-
      CreateDocumentRequest registers a document (metadata first; the bytes are
      uploaded beforehand and referenced by content_blob_id). When the content is
      already held by a live document, that document is reused instead (see
      CreateDocumentResponse.reused).
    required:
      - documentTypeCode
      - title
  v1CreateDocumentResponse:
    type: object
    properties:
      document:
        $ref: '#/definitions/v1Document'
        description: document is the created document, or the reused one when reused is true.
      createdEvent:
        $ref: '#/definitions/v1AuditEvent'
        description: created_event is DOCUMENT_CREATED, or DOCUMENT_REUSED on the existing document.
      initialRelationship:
        $ref: '#/definitions/v1SubjectRelationship'
        description: |-
          initial_relationship is the CASE_HAS_DOCUMENT edge; set only when
          link_to_case_id was provided and the link did not already exist.
      reused:
        type: boolean
        description: |-
          reused is true when the content was already held by a live document, which
          was reused (no new document; the request's title and metadata are ignored).
    description: CreateDocumentResponse returns the created (or reused) document and its side effects.
...
  v1Document:
    type: object
    properties:
      subjectRef:
        $ref: '#/definitions/v1SubjectRef'
        description: subject_ref is the embedded canonical identity (kind = DOCUMENT); its id is the document id.
      documentType:
        $ref: '#/definitions/v1DocumentType'
        description: document_type is the classification chosen at creation.
      title:
        type: string
        description: title is the document title (1-500 characters), mirrored into the subject label.
      description:
        type: string
        description: description is free text (at most 4000 characters).
      officialDate:
        type: string
        description: |-
          official_date is the legal probative date of the document itself, as
          YYYY-MM-DD; empty when unknown.
      externalSystem:
        type: string
        description: |-
          external_system names the system of record: alfresco, minio, sharepoint
          or goeland-legacy; empty when internal.
      externalId:
        type: string
        description: external_id is the authoritative id in external_system (prevents duplication).
      externalUrl:
        type: string
        description: external_url is a clickable link to the external document, if available.
      language:
        type: string
        description: |-
          language is the ISO 639 content language, e.g. "fr" or "de" (at most 10
          characters); unrelated to the UI locale.
      status:
        $ref: '#/definitions/v1DocumentStatus'
        description: status is the lifecycle state; FINAL when the current version is final.
      metadata:
        type: object
        description: 'metadata holds secondary data: classification, keywords, extracted entities...'
      createdAt:
        type: string
        format: date-time
        description: created_at is the server creation time.
        readOnly: true
      createdBy:
        type: string
        description: created_by is the operator who created the document.
      updatedAt:
        type: string
        format: date-time
        description: updated_at is the last modification time, maintained by the database.
        readOnly: true
      recordMetadata:
        $ref: '#/definitions/v1RecordMetadata'
        description: record_metadata is the governance record (confidentiality, lock, owner...).
      currentVersion:
        $ref: '#/definitions/v1DocumentVersion'
        description: current_version is the explicit current version, with its content.
        readOnly: true
    description: |-
      Document is the logical business document, 1:1 with SubjectRef (id == subject
      id); its content lives in versions. External references (Alfresco node, legacy,
      SharePoint) stay document-level.
    required:
      - title
...
  v1DocumentType:
    type: object
    properties:
      id:
        type: string
        description: id is the catalogue row UUID.
      code:
        type: string
        description: |-
          code is the unique stable key (1-100 characters), e.g. PLAN,
          INCOMING_LETTER or DECISION.
      label:
        type: string
        description: label is the human label (at most 200 characters).
      description:
        type: string
        description: description documents the business meaning (at most 2000 characters).
      category:
        type: string
        description: |-
          category groups types: ENTREE, SORTIE, PLAN, DECISION, JUSTIFICATIF...;
          empty when uncategorized.
      isActive:
        type: boolean
        description: is_active reports whether the type is offered for new documents.
    description: |-
      DocumentType is a controlled classification (behavior + category).
      Types are seeded reference data; a document's type is fixed at creation.
    required:
      - code
...
  v1ListDocumentVersionsResponse:
    type: object
    properties:
      versions:
        type: array
        items:
          type: object
          $ref: '#/definitions/v1DocumentVersion'
        description: versions are the versions with their content, newest first.
    description: ListDocumentVersionsResponse returns every version, newest first.
...
  v1SubjectRelationship:
    type: object
    properties:
      id:
        type: string
        description: id is the server-assigned edge UUID.
      source:
        $ref: '#/definitions/v1SubjectRef'
        description: source is the subject the edge starts from.
      target:
        $ref: '#/definitions/v1SubjectRef'
        description: target is the subject the edge points to.
      relationshipType:
        $ref: '#/definitions/v1RelationshipType'
        description: relationship_type is the validated type of the edge.
      roleDetail:
        type: string
        description: |-
          role_detail qualifies the relation in free text (at most 500 characters),
          e.g. "mandataire principal" or "plan de masse v2".
      validFrom:
        type: string
        format: date-time
        description: valid_from is the optional business start of validity.
      validTo:
        type: string
        format: date-time
        description: |-
          valid_to is the business end of validity, set by EndRelationship; absent
          while the edge is open.
      createdAt:
        type: string
        format: date-time
        description: created_at is the server creation time.
        readOnly: true
      createdBy:
        type: string
        description: created_by is the operator who created the edge.
      deletedAt:
        type: string
        format: date-time
        description: deleted_at marks an unlinked (soft-deleted) edge; absent while active.
        readOnly: true
    description: |-
      SubjectRelationship is an actual edge in the business graph: typed, validated, soft-deletable.
      At most one open edge (not unlinked, no valid_to) exists per (source, target, type);
      an ended edge stays as history.
...